+++ src/asynchron/codegen/spec/reader/json.py
@@ -6,7 +6,7 @@ import json
 import typing as t
 
 from asynchron.codegen.app import AsyncApiConfigReader
-from asynchron.codegen.spec.base import AsyncAPIObject
+from asynchron.codegen.spec.asyncapi import AsyncAPIObject
 
 
 # TODO: handle IO, yaml and pydantic errors
